+++ bricknil/const.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Useful constants for BlueBrick
+"""Useful constants for BrickNil
 """
 from enum import Enum
 import platform